+++ src/library/xevents.h
@@ -32,9 +32,6 @@ namespace libtas {
 void pushNativeXlibEvents(void);
 void pushNativeXlibEvents(Display *display);
 
-/* Wait for the xevent queue to become empty (returns true) or timeout (returns false) */
-bool syncXEvents();
-
 OVERRIDE int XNextEvent(Display *display, XEvent *event_return);
 OVERRIDE int XPeekEvent(Display *display, XEvent *event_return);
 OVERRIDE int XWindowEvent(Display *display, Window w, long event_mask, XEvent *event_return);